findbugs RV_ABSOLUTE_VALUE_OF_HASHCODE.  Codes should cast to long before the abs operation.
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@23550 0c594bb0-8543-0410-a25e-d8ec440261db

+++ Autoplot/src/org/autoplot/JythonUtil.java
@@ -232,7 +232,7 @@ public class JythonUtil {
         String okayedContents= okayed.get(filename);
         if ( okayedContents==null ) {
             final File lastVersionDir= Paths.get( AutoplotSettings.settings().resolveProperty( AutoplotSettings.PROP_AUTOPLOTDATA ), "scripts" ).toFile();
-            final File lastVersionFile= Paths.get( lastVersionDir.toString(), String.format( "%010d.jy", Math.abs(filename.hashCode()) ).trim() ).toFile();        
+            final File lastVersionFile= Paths.get( lastVersionDir.toString(), String.format( "%010d.jy", Math.abs( (long)filename.hashCode()) ).trim() ).toFile();        
             if ( lastVersionFile.exists() ) {
                 try {
                     String lastVersionContents= FileUtil.readFileToString(lastVersionFile);
@@ -591,7 +591,7 @@ public class JythonUtil {
                                 }
                             }
                             
-                            final File lastVersionFile= Paths.get( lastVersionDir.toString(), String.format( "%010d.jy", Math.abs(file.toString().hashCode()) ).trim() ).toFile();        
+                            final File lastVersionFile= Paths.get( lastVersionDir.toString(), String.format( "%010d.jy", Math.abs((long)file.toString().hashCode()) ).trim() ).toFile();        
                             
                             FileUtil.fileCopy( file, lastVersionFile );
